backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
Flask_WTF==1.1.1
firebase-admin == 6.1.0
googlemaps==4.7.3
numpy==1.24.1
Pillow==9.4.0
pyrebase==3.0.27
Pyrebase4==4.6.0
python-dotenv==1.0.0
requests==2.28.1
Werkzeug==2.2.2
WTForms==3.0.1
